Upgrade to Pex 2.1.61 and use `--venv` copies. (#13979)
This picks up a fix for `--venv` copies (hardlinks) mode that allows
Pants to work with more distributions (including its own!).

See the changelog here:
  https://github.com/pantsbuild/pex/releases/tag/v2.1.61

An advanced `pex` option is plumbed to use symlinks when possible, but
but does not turn this on by default since it can break given certain
distributions. Pants does dogfood the optimization though since it is
not plagued by such distributions.