[internal] Make more explicit module mapping code for Python, Shell, and Protobuf (#13981)
Because we know that a `SingleSourceField` has exactly one file—and we can ban `*` so can compute its file path without using the engine—we can simplify dependency inference when creating the global module mapping. There's no need to actually look up the file system.

A downside of this PR is that we have lazier validation that the `source` field matches files on disk, i.e. that the file is there. But there are still plenty of places this gets validated, such as when computing the dependencies of a target because we have to inspect its source code. (Compared to creating a global module mapping for dep inference in general.)

There's a marginal speed up, too:

Before:

```
❯ hyperfine -w 1 -r 25 './pants --no-pantsd dependencies src/python/pants/util/strutil.py:util'
Benchmark #1: ./pants --no-pantsd dependencies src/python/pants/util/strutil.py:util
  Time (mean ± σ):      4.576 s ±  0.139 s    [User: 3.180 s, System: 0.657 s]
  Range (min … max):    4.373 s …  5.003 s    25 runs
```

After:

```
❯ hyperfine -w 1 -r 25 './pants --no-pantsd dependencies src/python/pants/util/strutil.py:util'
Benchmark #1: ./pants --no-pantsd dependencies src/python/pants/util/strutil.py:util
  Time (mean ± σ):      4.487 s ±  0.172 s    [User: 3.019 s, System: 0.628 s]
  Range (min … max):    4.224 s …  5.094 s    25 runs
```

[ci skip-rust]
[ci skip-build-wheels]