[internal] Add `experimental_compatible_resolves` to `python_requirement` and use it to generate lockfiles (#13978)
Part of the design from https://github.com/pantsbuild/pants/issues/13621. This change has worked well for JVM.

This does have an open question though of how we should determine which interpreter constraints to use for each resolve. `python_requirement` targets do not have constraints themselves, so we either need to associate each resolve with particular constraints via global config, or we need to look at all consumers of the resolves to calculate it. For now, we just use the global default.

Reminder that these changes are only enabled if you use `--python-enable-resolves`.

[ci skip-rust]
[ci skip-build-wheels]