[internal] Change contributors script to not cite # of contributions (#14005)
We decided as maintainers that we don't want to emphasize the # of commits in our weekly release emails thanking code contributors. # of commits is not a great way to measure contributions, and it also creates a hierarchy we don't want to emphasize.

Instead, the maintainers informally voted to drop # of contributions and sort the contributors alphabetically. We also considered randomly.

In the process, this rewrites the contributor script to use Python.

[ci skip-rust]
[ci skip-build-wheels]