Add new `image_labels` field to `docker_image`. (#13973)
This builds on top of previous PRs, one commit per PR. May review just the last commit, for this PR (or postpone review until previous PRs has landed)

```python
docker_image(
  ...
  image_labels={
    "my.label": "some value",
  }
)
```

### Use case

Using the `LABEL` Dockerfile instruction gets you a long way. However, when the `docker_image` target is defined using a macro, and the labels you want to define depends on inputs on that macro call, it becomes cumbersome and ugly to rely on `LABEL`s in the Dockerfile.