[jvm] Parse maven coordinates using regular expression (cherrypick of #14010) (#14015)
This goes on top of #13996 removing the workaround on the maven coordinates after finding out that coursier's behaviour (reported at coursier/coursier#2315) is the correct one. While I haven't found any _spec_ of that, this is what I gather from [this Maven document](https://maven.apache.org/plugins/maven-assembly-plugin/examples/single/including-and-excluding-artifacts.html) and from inspecting other implementations.

This PR parses the Maven coordinates using a regular expression that is able to extract the packaging and classifier in case they are present. This should make it compatible with artifacts that only display three of the elements and with the ones that even include the classifier in it.

[ci skip-rust]
[ci skip-build-wheels]