[internal] Add `update-build-files` fixer for changing Python CAOFs to target generators (#14039)
Part of https://github.com/pantsbuild/pants/issues/12915.

Changing `python_requirements` and friends to be a target generator, rather than context-aware object factory (CAOF), is particularly tricky because all references must be updated from `dir:req` to `dir#req`. Even more complex, the target generator will sometimes need a `name` so it becomes `dir:reqs#req`. That's far too hard and error-prone for users to figure out.

This fixer updates all _references_ to generated targets. 

It does not automatically add the `name=` when necessary, though, because that's hard to pull off with updating the BUILD file - instead we log an error warning with what to do:

```
18:42:55.79 [ERROR] You must manually add the `name=` field to the following targets. This is not done automatically by the `update-build-files` goal.

  * `python_requirements` in 3rdparty/python/BUILD: add `name="reqs"
```

In a followup, we'll add an option to toggle whether the symbol `python_requirements` et al should refer to the CAOF vs. target generator. This new fixer will only be used if CAOFs are in use (the default).

[ci skip-rust]