`_python_requirements_file` target uses `source: str` field instead of `sources: list[str]` (#14082)
To minimize behavior changes while implementing https://github.com/pantsbuild/pants/pull/14075, we want to make sure that `--changed-since=HEAD --changed-dependees=direct` will still claim that all generated `python_requirement` targets are impacted if the `requirements.txt` changed.

This would happen naturally if a generated target depended on its target generator: https://github.com/pantsbuild/pants/issues/13118. But that will require a lot more design work and is a big change we're not ready to make.

So we can work around this by still having `_python_requirements_file` even with target generators. But before doing that, we should change to `source: str` to actually represent the expected API of the target.

Because this is a private API, we make a breaking change.

[ci skip-rust]
[ci skip-build-wheels]