[internal] Add & Use flake8-no-implicit-concat
Sometimes the stuff on HN is useful:
https://news.ycombinator.com/item?id=29841560
The pants repo doesn't actually have bugs like this in our repo. I did fix a couple of places, see https://github.com/pantsbuild/pants/pull/14119, but those are not bugs.

But this is something that could happen downthe line, so it is a good idea to have this check enabled.

https://github.com/10sr/flake8-no-implicit-concat