Adds `.env` file to make vscode auto-imports work correctly out of the box (#14130)
Per the docs, vs code's Python source roots are defined in a .env file in the root of a repo. If such a file does not exist, it makes guesses that are not correct for the Pants repo. When this .env file is in place, auto-import works fine out of the box. The result of this is not needing to correct auto-generated imports nearly as often (normally I'd discover these incorrect imports deep during a pants run). It doesn't seem to be harmful to any of the tools that I otherwise use in the Pants repo.