Add `#pants: ignore` pragma for ignoring imports (#14059)
Adding a way to tell pants __not__ to infer what is seemingly a very legitimate dependency.

See thread: https://pantsbuild.slack.com/archives/C0D7TNJHL/p1641227325498500

For documentation purposes: `# pants: ignore` suppresses the inference of an import for all imports inferred for the line it is on.

N.B. There isn't way to have a single `# pants: ignore` suppress multi lines (e.g. putting it on the open or closing parenthesis does something special). This is shown in the tests.