Deprecate `python_requirements` and `poetry_requirements` using old macro in favor of target generation (#14075)
Closes https://github.com/pantsbuild/pants/issues/12915.

## Benefits of switching to target generation

* Closes https://github.com/pantsbuild/pants/issues/7022.
* Target generators show up in `./pants help`.
* Consistent mechanism for target generation, as explained at https://www.pantsbuild.org/docs/targets.

## Deprecation plan

This PR adds a new global option `--use-deprecated-python-macros` with a default of `True`. When True, `parser.py` is taught to load the old macros; else it loads the new targets.

The PR deprecates the default of `--use-deprecated-python-macros` so that it will default to `False` in Pants 2.11. Then, we'll deprecate the option outright and remove it in Pants 2.12, meaning the macros will be removed then too.

## How users upgrade

```
❯ ./pants
17:06:21.30 [WARN] DEPRECATED: the option `--use-deprecated-python-macros` defaulting to true will be removed in version 2.11.0.dev0.

In Pants 2.11, the default for the global option `--use-deprecated-python-macros` will change to false.

To fix this deprecation, explicitly set `use_deprecated_python_macros = true` in the `[GLOBAL]` section of `pants.toml`. Or, When you are ready to upgrade to the improved target generation mechanism, follow these steps:

  1. Run `./pants update-build-files --fix-python-macros`
  2. Check the logs for an ERROR log to see if you have to manually add `name=` anywhere.
  3. Set `use_deprecated_python_macros = false` in `[GLOBAL]` in pants.toml.

(Why upgrade from the old macro mechanism to target generation? Among other benefits, it makes sure that the Pants daemon is properly invalidated when you change `requirements.txt` and `pyproject.toml`.)
```