Deprecate `sourcefile-validation` in favor of `regex-lint` (#14173)
This checker will soon be moved from the `validate` goal to the `lint` goal. When we do that, we need to have a name for the linter, e.g. in this message:

```
flake8 succeeded.
isort failed.
regex-lint succeeded.
```

We bike-shedded in Slack and came up with `regex-lint`:

- `lint` is better than `validate` because this is the `lint` goal
- `regex` is because the user has to provide regex patterns

We rejected the similar `regex-linter` because that sounds like it is linting regex expressions themselves.

[ci skip-rust]