[internal] go: rewrite third party package analysis to not use `go list` (#14157)
## Problem

The existing Go backend uses `go list` to analyze third party modules. Several users encountered a problem where the call to `go list` was demanding changes to `go.mod` with this error message:

```
go: updates to go.mod needed; to update it:
	go mod tidy
```

Running `go mod tidy`, however, made no changes to the `go.mod`, so there was no way to ever satisfy `go list`.

## Solution

Solve the issue by rewriting the Go backend to use the same custom package analyzer used for first-party package analysis.  This removes the reliance on `go list` for package analysis which was the `go` invocation that was generating the error message. (`go list` is still used to compute what modules are dependencies.)

`ThirdPartyPkgInfo` is renamed to `ThirdPartyPkgAnalysis` to better align with `FirstPartyPkgAnalysis`.

[ci skip-rust]