[internal] Inline `ValidatedJvmToolLockfileRequest` into `ToolClasspathRequest` (#14241)
To consume a tool lockfile, we before would:

1. Read the lockfile, either by:
     1. Using `ValidatedJvmToolLockfileRequest`, which only works if it's a `JvmToolBase`
     2. Using `GenerateJvmLockfile -> CoursierResolvedLockfile` rule, which we used with `scalac_plugins.py`
2. Pass the lockfile to `ToolClasspathRequest -> ToolClasspath`

The rule in 1.2 was not using the validation added in https://github.com/pantsbuild/pants/pull/14185, and it was generally confusing to have two code paths.

This unifies everything by having `ToolClasspathRequest -> ToolClasspath` now handle reading, validating, and then fetching the lockfile. 

Reading and validating a lockfile requires the same metadata we need for generating a lockfile with `generate-lockfiles` via `GenerateJvmLockfile`. So, we reuse that type for the sake of DRY.

[ci skip-rust]