[internal] Add `[python].experimental_resolves_to_interpreter_constraints` (#14254)
## Problem

We need to know which interpreter constraints to use when generating a user lockfile, which will affect the behavior of Poetry & (soon) Pex.

The interpreter constraints of the resolve itself should be a superset of the user's actual code consuming that resolve. For example, the resolve might work with 3.6-3.9, and then there are some individual `python_test` targets that are specifically 3.6, 3.7, etc. It's necessary that the resolve can accommodate those more specific values.

## Solution

Let the user manually set the interpreter constraints for each resolve, with it defaulting to the global interpreter constraints.

This setting only impacts lockfile generation—the value is not mixed into call sites' computation of which interpreter constraints to use for user code. However, we store the resolve's interpreter constraints in the lockfile's header, which allows us to validate that the user's code is a subset of the resolve's constraints.

## Alternatives rejected

### Magically determine ICs based on consumers

We would scan all consumers of a resolve, then merge together to form interpreter constraints. 

@stuhood and I believe this is too magical and brittle, that you could too easily inadvertently change the ICs of an entire resolve by changing the `interpreter_constraints` of one target.

### Add `interpreter_constraints` to `python_requirement` target

Each individual target would have the field, and we would merge all `python_requirement` targets for that resolve. 

I believe this is too fine-grained and thus confusing for users. It would be tedious to have to set the interpreter constraints for each individual requirement, and it's also hard to be accurate because you would have to reason about transitive deps which are not pinned until you generate the lockfile, so could change when regenerating. Instead, treat each resolve as an atomic unit.

By avoiding this design, we also avoid having the resolve's interpreter constraints changing the code's computed interpreter constraints, outside of lockfile validation. Everything behaves the same as before.

[ci skip-rust]