[internal] `python_requirement` target generators pass down some fields to generated targets (#14273)
[ci skip-rust]
Closes https://github.com/pantsbuild/pants/issues/14268.

Now it's ergonomic to set `tags`, `description`, and `compatible_resolves` on all generated targets. As before, you can also use `overrides` for per-target settings. With `compatible_resolves`, you can also leave it off and use `[python].default_resolve`.

This does mean that it's impossible to have `tags` and `description` _just_ on the target generator and not generated targets. But that is already the case with `python_sources -> python_source` etc. So this makes things consistent, at least.

[ci skip-rust]
[ci skip-build-wheels]