Add `setrlimit` filter to coursier invocations (#14296)
After a few attempts to close #13942, I have emerged from a giant pile of yaks and discovered that the solution was actually attached to a sheep.

Annoyingly, a large number of our coursier invocations are used to look at outputs from stderr, which makes using `grep` (as suggested as an easy fix in #13942) to filter out spurious warnings quite difficult. Instead, this change adds another wrapper in the form of a Python script, which preserves separation of stdout and stderr for all Coursier invocations, but removes all instances of the setrlimit warning from stderr before outputting it. 

I've also renamed `coursier_wrapper`* to `coursier_fetch_wrapper`* to indicate that only `coursier fetch` calls are wrapped; calls like `coursier java-home` continue to be called more directly.