Don't render skipped tools in final summary with `fmt`, `lint`, `check`, and `test` (#14303)
Before:

```
❯ ./pants --black-skip fmt src/python/pants/util/strutil.py
07:22:21.51 [INFO] Completed: Format with Autoflake - autoflake made no changes.
07:22:22.52 [INFO] Completed: Format with docformatter - Docformatter made no changes.
07:22:22.54 [INFO] Completed: Format with isort - isort made no changes.

- Black skipped.
✓ Docformatter made no changes.
✓ autoflake made no changes.
✓ isort made no changes.
```


After:

```
❯ ./pants --black-skip fmt src/python/pants/util/strutil.py                                           
07:22:01.24 [INFO] Completed: Format with Autoflake - autoflake made no changes.
07:22:01.41 [INFO] Completed: Format with docformatter - Docformatter made no changes.
07:22:01.70 [INFO] Completed: Format with isort - isort made no changes.

✓ Docformatter made no changes.
✓ autoflake made no changes.
✓ isort made no changes.
```

As we've added more tools, rendering skipping is somewhat noisy. 

More importantly, it causes a problem when tools want to disable themselves via some complex logic that requires the Rules API. For example, with Go, we don't know if a `go_package` has tests in it or not until compiling the code and running our analyzer: 

https://github.com/pantsbuild/pants/blob/0488605f54ef3602797d68a5c4797dea9d706b14/src/python/pants/backend/go/goals/test.py#L192-L193

With the upcoming `regex-lint` (formerly `validate`), we want to only enable the linter if `[regex-lint].config` is set up: https://github.com/pantsbuild/pants/pull/14102. 

In both these cases, it is annoying & noisy to render the skipped tools in the summary. I wasn't anticipating these use cases when designing this summary in 2020: https://github.com/pantsbuild/pants/pull/9710.

We will still log that the tool is skipped at `-ldebug`, which helps if users are confused why something isn't showing up.

[ci skip-rust]
[ci skip-build-wheels]