`LintRequest`, `FmtRequest`, and `CheckRequest` must set `name: ClassVar[str]` (#14304)
This will be necessary for us to add `lint --only=isort` and so on. To do that, we need to know _before_ executing any requests what the name of each tool is, so that we can then match it against 

I at first called this `tool_request`, but realized a plugin might not always be a "tool", like `regex-lint` (aka `validate`) which is in-memory. Simply `name` seems the most flexible.

I originally also renamed `FmtResult(formatter_name=)` to be `name=` for parity, but I reverted it to reduce churn for Plugin developers. That is still probably a good change to do, but we could do it more gracefully with a deprecation warning. It's low priority either way.

[ci skip-rust]