`ReplImplementation` now passes root targets, not transitive closure (#14323)
It is useful for implementations to distinguish between what were the input targets, such as for https://github.com/pantsbuild/pants/issues/14295.

This also facilitates a proposal I made in Slack. Default to `repl --shell=<infer>`, where we inspect the _root_ targets (not transitive closure) to see what target type is most common. If Python, use Python. And so on. 

[ci skip-rust]
[ci skip-build-wheels]