Rename Python resolve options and fields to not say "experimental" (#14326)
We believe that this feature is now stable enough to not hedge with the `--experimental` name. The key insight is that we won't have `compatible_resolves` on `python_source` targets, per https://github.com/pantsbuild/pants/pull/14299. 

We still don't have this feature on by default via `--enable-resolves`, and we likely won't until using PEX for lockfile generation. But this lowers the barrier to entry for people who want to try out this new mechanism.

[ci skip-rust]