Rename the `download-pex-bin` scope to `pex-cli`.
This is for uniformity with other tool config scopes, which are named for
the tool and not for the act of downloading them. I'd have liked to use
just `pex` but that is used to configure pex invocations.

While here, standardized terminology in this file, to avoid overloading
the term "pex_binary", which is already a target type.

[ci skip-rust]

[ci skip-build-wheels]