golang: add support for Go 1.18 fuzz targets to `go test` (#14357)
Go 1.18 adds support for ["fuzz tests"](https://go.dev/doc/fuzz/). The addition of this support changes the API exported by the `testing` package for test runners which will break `./pants test` on Go 1.18 if Pants were not to adapt to this change.

Also, modify the template for the generated `testmain.go` to clean up extra whitespace in the generated code. The code now looks nicer when printed out for debugging.

Closes https://github.com/pantsbuild/pants/issues/13691.

[ci skip-rust]

[ci skip-build-wheels]