Even better redesign of MyPy-friendly options registration (#14331)
Adding `option_types.py` which is meant to be a complete set of all possible option types. Each possible type has been stamped out into a concretely(-ish for Enums) typed class.

__Great__ (great, great, _great_) care has been taken to make this as boilerplate-free and "automagical" as possible. Unfortunately, after countless hours, I've come to realize there is no no-boilerplate solution. This one at least contains the boilerplate to one file, so that's ultimately a huge win.

The client code retrofits will follow in a separate PR. All the Pants call sites have been audited to ensure the kwargs are handled here.

I don't expect this to be the final iteration of this code. But I wanted to lay enough groundwork that 80% of the existing code can switch over without a hitch. Further modifications will possibly be made in the same PR as switchover.

![oh yeah](https://media.giphy.com/media/KEYEpIngcmXlHetDqz/giphy.gif)