Adds `run` support for `deploy_jar` targets (#14352)
This adds `run` support for `deploy_jar` targets. It works! Closes #14283, at least until `war` files are properly supported.

To make Scala binaries work, we now inject a dependency for the version of `scala-library` specified by `--scala-version` at compile time, whenever `scala-library` is not provided as a transitive dependency. This is a good-enough stop-gap in service of #14171.

There's one ugly caveat which I'm reserving for future work:

For tool support, Couriser downloads a JVM into an append-only cache for most JVM processes. Since `InteractiveProcess` does not support these, I've added `RuntimeJdk__`, which downloads a JDK into a `Digest`. The digest gets written into the temporary directory alongside the runnable artifacts. It's not ideal, but for now, it removes the JDK download process from the `run` process, and the JDK is only downloaded once per `pantsd` execution. It's probably _ok_ for now. Being able to reuse the global append-only cache directories will be a better solution once we get there.