[internal] Extracts `JdkEnvironment` from global `JdkSetup` object (#14400)
This is pre-work for #13995, which will require us to explicitly provide a JDK environment to `JvmProcess`es.

The substantive changes are in `jdk_rules.py` -- `JdkSetup` now becomes `JdkEnvironment` (constructed by `JdkRequest`), and the new `JdkSetup` global object is just a simple wrapper for `JdkEnvironment` based on the provided `--jvm-jdk` option.

Everything else is adding `JdkSetup` injections back to our JDK-consuming rules. Alas.