scala: support multiple scala versions (#14425)
Implement support for multiple Scala versions in the same repository. Each resolve will be assigned a single Scala version.

Scala requires that all Scala-compiled jars on a classpath (i.e, a resolve) be "compatible" with one another (regardless of whether they are third-party jars or jars compiled from first-party sources). For Scala 2.x versions, for example, "compatibility" means that all Scala-compiled jars were compiled with the same minor version of Scala (i.e., 2.12). We can maintain this "Scala compatibility" property for a resolve by assigning the resolve a single Scala version which ensures that there is single Scala version that all of the third-party jars must be compatible with and that all first-party sources are compiled with. 

Introduce the `--scala-version-for-resolve` option to implement the mapping from resolve to Scala version. The `--scala-version` option is deprecated and will be used as a fallback if the Scala version for a particular resolve is not set in `--scala-version-for-resolve`.

Closes https://github.com/pantsbuild/pants/issues/13994.