More declarative target generators for plugins (#14377)
#13882 will require adjusting how many instances of a target are created based on how many times the `parametrize` keyword appears in its fields (including in the `overrides` field, in the case of target generators). The `parametrize` keyword is not itself a valid field value (and would also be challenging to render as such in `peek`), and so target generation must be updated to consume something other than the instantiated target generator to decide which fields to generate with.

To prepare for these changes, the base ("`template`") and overridden fields to use in target generation are added to the `GenerateTargetsRequest` class, and made required in `generate_file_level_targets`. To reduce the total amount of churn in callers of `generate_file_level_targets`, this change introduces the `TargetGenerator` and `TargetFilesGenerator` base classes for `Target`. The latter allows file-target generation to be requested declaratively, and both allow specifying which fields should be "moved" and "copied" from the target generator to the generated targets. These moved/copied declarations are used to construct the `template` for generated targets.

[ci skip-rust]
[ci skip-build-wheels]